backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Scheduler
apscheduler==3.10.4

# Crawler Dependencies
playwright==1.41.0
pyyaml==6.0.1

# Knowledge Graph Extraction
openai==1.10.0
tiktoken==0.5.2
zhipuai==2.0.1

# Neo4j
neo4j==5.15.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
